Refactor client_db.cc into clientdb/libclientdb.la
- Remove unused clientdbSetWriteLimiter
- Replace hash_link with std::map storage
- separate ClientInfo into own files
- separate SNMP pieces into conditional file
