+++ configure.ac
@@ -3682,6 +3682,7 @@ AC_CONFIG_FILES([
 	src/auth/ntlm/SMB_LM/Makefile
 	src/auth/ntlm/SSPI/Makefile
 	src/base/Makefile
+	src/clientdb/Makefile
 	src/clients/Makefile
 	src/comm/Makefile
 	src/dns/Makefile
