+++ src/clientdb/ClientInfo.h
