+++ src/clientdb/Makefile.am
@@ -1,0 +1,21 @@
+## Copyright (C) 1996-2021 The Squid Software Foundation and contributors
+##
+## Squid software is distributed under GPLv2+ license and includes
+## contributions from numerous individuals and organizations.
+## Please see the COPYING and CONTRIBUTORS files for details.
+##
+
+include $(top_srcdir)/src/Common.am
+include $(top_srcdir)/src/TestHeaders.am
+
+noinst_LTLIBRARIES = libclientdb.la
+
+libclientdb_la_SOURCES= \
+	Cache.cc \
+	Cache.h \
+	ClientInfo.cc \
+	ClientInfo.h \
+	forward.h \
+	Runner.cc \
+	SnmpGadgets.cc \
+	SnmpGadgets.h
