+++ src/clientdb/forward.h
@@ -1,0 +1,22 @@
+/*
+ * Copyright (C) 1996-2021 The Squid Software Foundation and contributors
+ *
+ * Squid software is distributed under GPLv2+ license and includes
+ * contributions from numerous individuals and organizations.
+ * Please see the COPYING and CONTRIBUTORS files for details.
+ */
+
+/* DEBUG: section 77    Client Database */
+
+#ifndef _SQUID__SRC_CLIENTDB_FORWARD_H
+#define _SQUID__SRC_CLIENTDB_FORWARD_H
+
+class ClientInfo;
+
+/// Database of client information
+namespace ClientDb {
+
+
+} // namespace ClientDb
+
+#endif /* _SQUID__SRC_CLIENTDB_FORWARD_H */
