+++ src/comm.cc
@@ -9,7 +9,7 @@
 /* DEBUG: section 05    Socket Functions */
 
 #include "squid.h"
-#include "ClientInfo.h"
+#include "clientdb/ClientInfo.h"
 #include "comm/AcceptLimiter.h"
 #include "comm/comm_internal.h"
 #include "comm/Connection.h"
