+++ src/tests/Stub.am
@@ -15,7 +15,6 @@ STUB_SOURCE = \
     tests/stub_CachePeer.cc \
     tests/stub_carp.cc \
     tests/stub_cbdata.cc \
-    tests/stub_client_db.cc \
     tests/stub_client_side.cc \
     tests/stub_client_side_request.cc \
     tests/stub_CollapsedForwarding.cc \
@@ -49,6 +48,7 @@ STUB_SOURCE = \
     tests/stub_libanyp.cc \
     tests/stub_libauth_acls.cc \
     tests/stub_libauth.cc \
+    tests/stub_libclientdb.cc \
     tests/stub_libcomm.cc \
     tests/stub_libdiskio.cc \
     tests/stub_liberror.cc \
